Move some eqK genK
