+++ arrow-fx-test/src/main/kotlin/arrow/fx/test/eq/EqK.kt
@@ -2,8 +2,10 @@ package arrow.fx.test.eq
 
 import arrow.Kind
 import arrow.core.extensions.either.eq.eq
+import arrow.fx.DecisionPartialOf
 import arrow.fx.ForIO
 import arrow.fx.IO
+import arrow.fx.Schedule
 import arrow.fx.extensions.io.applicative.applicative
 import arrow.fx.extensions.io.applicativeError.attempt
 import arrow.fx.extensions.io.concurrent.waitFor
@@ -50,4 +52,14 @@ fun Fiber.Companion.eqK() = object : EqK<FiberPartialOf<ForIO>> {
  */
 fun throwableEq() = Eq { a: Throwable, b ->
   a::class == b::class && a.message == b.message
+}
+
+fun Schedule.Decision.Companion.eqK(): EqK<DecisionPartialOf<Any?>> = object : EqK<DecisionPartialOf<Any?>> {
+  override fun <A> Kind<DecisionPartialOf<Any?>, A>.eqK(other: Kind<DecisionPartialOf<Any?>, A>, EQ: Eq<A>): Boolean =
+    (fix() to other.fix()).let { (l, r) ->
+      l.cont == r.cont &&
+        l.delay.nanoseconds == r.delay.nanoseconds &&
+        // don't force end result if we don't continue as it may contain Nothing
+        (if (l.cont) EQ.run { l.finish.value().eqv(r.finish.value()) } else true)
+    }
 }
